# Avoid allocation during validation