--- a/build.gradle
+++ b/build.gradle
@@ -42,6 +42,7 @@ dependencies {
     implementation 'com.fasterxml.jackson.core:jackson-annotations'
 
     testImplementation 'com.fasterxml.jackson.core:jackson-databind'
+    testImplementation 'net.jqwik:jqwik'
     testImplementation 'org.junit.jupiter:junit-jupiter'
 
     jmh 'org.openjdk.jmh:jmh-core'
@@ -51,6 +52,12 @@ dependencies {
 
 configurations.testCompileClasspath {
     exclude module: 'junit'
+}
+
+tasks.withType(Test) {
+    useJUnitPlatform {
+        includeEngines 'jqwik', 'junit-jupiter'
+    }
 }
 
 jmh {