--- a/changelog/@unreleased/pr-140.v2.yml
+++ b/changelog/@unreleased/pr-140.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: Validating ResourceIdentifier components no longer allocates.
+  links:
+  - https://github.com/palantir/resource-identifier/pull/140