--- a/src/test/java/com/palantir/ri/ResourceIdentifierTest.java
+++ b/src/test/java/com/palantir/ri/ResourceIdentifierTest.java
@@ -99,6 +99,12 @@ final class ResourceIdentifierTest {
     @Test
     void testConstructionErrorMessage() {
         try {
+            ResourceIdentifier.of(null);
+            fail();
+        } catch (IllegalArgumentException e) {
+            assertEquals("Illegal resource identifier format: null", e.getMessage());
+        }
+        try {
             ResourceIdentifier.of("ri.bad....dots");
             fail();
         } catch (IllegalArgumentException e) {