--- a/versions.props
+++ b/versions.props
@@ -1,3 +1,4 @@
 com.fasterxml.jackson.core:* = 2.11.3
+net.jqwik:* = 1.5.3
 org.junit.jupiter:* = 5.7.2
 org.openjdk.jmh:* = 1.32