… src/test/java/com/palantir/ri/ResourceIdentifierPropertyTest.java
… versions.lock