Implementation of suggestion from https://github.com/palantir/resource-identifier/pull/139#issuecomment-884425907.

The methods to create and validate RIDs use regular expressions to validation the RID components. This can result in a non-trivial number of allocations for applications that validate a lot of RIDs.

This PR replaces the regular expressions with a simple scan over the input, allowing us to perform these validations without allocating.