@carterkozak 2021-07-22
Any chance you can share before/after results from https://github.com/palantir/resource-identifier/blob/develop/src/jmh/java/com/palantir/ri/ResourceIdentifierBenchmark.java ? Might be helpful to add a second benchmark which creates RID instances using https://github.com/palantir/resource-identifier/blob/6a47ea8e2c2a70a82342a1728538ad0b0a863597/src/main/java/com/palantir/ri/ResourceIdentifier.java#L262