@carterkozak 2021-07-22
We should implement a jquik test against the original regex along these lines to ensure we're not introducing regressions:
https://github.com/palantir/sls-version-java/pull/455/files#diff-d90462e86f4e5e19c0df1b4545c7f9dcd7b440d389a391db7eca402d2f39b520