@pkoenig10 2021-07-22
Benchmark results look pretty good. Much faster and uses way less memory.

### Before
```
Benchmark                                                                    (input)  Mode  Cnt     Score     Error   Units
ResourceIdentifierBenchmark.fromString                                   NO_INSTANCE  avgt    3   629.426 ± 146.220   ns/op
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate                    NO_INSTANCE  avgt    3  1287.626 ± 304.080  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate.norm               NO_INSTANCE  avgt    3   248.029 ±   0.004    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space           NO_INSTANCE  avgt    3  1308.892 ± 650.673  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space.norm      NO_INSTANCE  avgt    3   252.211 ± 182.929    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space       NO_INSTANCE  avgt    3     0.008 ±   0.027  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space.norm  NO_INSTANCE  avgt    3     0.001 ±   0.005    B/op
ResourceIdentifierBenchmark.fromString:·gc.count                         NO_INSTANCE  avgt    3    62.000            counts
ResourceIdentifierBenchmark.fromString:·gc.time                          NO_INSTANCE  avgt    3    34.000                ms
ResourceIdentifierBenchmark.fromString                                          UUID  avgt    3   621.556 ± 223.368   ns/op
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate                           UUID  avgt    3  1303.689 ± 460.886  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate.norm                      UUID  avgt    3   248.035 ±   0.013    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space                  UUID  avgt    3  1311.384 ± 952.597  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space.norm             UUID  avgt    3   249.443 ± 105.543    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space              UUID  avgt    3     0.007 ±   0.034  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space.norm         UUID  avgt    3     0.001 ±   0.007    B/op
ResourceIdentifierBenchmark.fromString:·gc.count                                UUID  avgt    3    75.000            counts
ResourceIdentifierBenchmark.fromString:·gc.time                                 UUID  avgt    3    33.000                ms
```

### After
```
Benchmark                                                                    (input)  Mode  Cnt     Score     Error   Units
ResourceIdentifierBenchmark.fromString                                   NO_INSTANCE  avgt    3    56.810 ±   6.366   ns/op
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate                    NO_INSTANCE  avgt    3  1840.210 ± 207.393  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate.norm               NO_INSTANCE  avgt    3    32.003 ±   0.002    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space           NO_INSTANCE  avgt    3  1858.339 ± 828.022  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space.norm      NO_INSTANCE  avgt    3    32.318 ±  13.759    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space       NO_INSTANCE  avgt    3     0.009 ±   0.031  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space.norm  NO_INSTANCE  avgt    3    ≈ 10⁻⁴              B/op
ResourceIdentifierBenchmark.fromString:·gc.count                         NO_INSTANCE  avgt    3    73.000            counts
ResourceIdentifierBenchmark.fromString:·gc.time                          NO_INSTANCE  avgt    3    52.000                ms
ResourceIdentifierBenchmark.fromString                                          UUID  avgt    3    56.985 ±   4.200   ns/op
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate                           UUID  avgt    3  1834.340 ± 133.566  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.alloc.rate.norm                      UUID  avgt    3    32.003 ±   0.001    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space                  UUID  avgt    3  1856.929 ± 812.571  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Eden_Space.norm             UUID  avgt    3    32.396 ±  12.475    B/op
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space              UUID  avgt    3     0.012 ±   0.045  MB/sec
ResourceIdentifierBenchmark.fromString:·gc.churn.G1_Survivor_Space.norm         UUID  avgt    3    ≈ 10⁻⁴              B/op
ResourceIdentifierBenchmark.fromString:·gc.count                                UUID  avgt    3    73.000            counts
ResourceIdentifierBenchmark.fromString:·gc.time                                 UUID  avgt    3    51.000                ms
```