@svc-autorelease 2021-07-22
Released [1.4.0](https://github.com/palantir/resource-identifier/releases/tag/1.4.0)